Split PID class from the state_commanded source files
