+++ src/state_commanded.h
@@ -5,6 +5,7 @@
 #include <tick-timer.h>
 
 #include "roaster.h"
+#include "eeprom_settings.h"
 
 // delay between SSR and heat relay transitions
 #define CONTROL_HEAT_SSR_RELAY_DELAY_MS 11U
@@ -106,40 +107,6 @@ class ControlDrum : public ControlPWM {
         void         _abortAction();
 };
 #endif // USE_STEPPER_DRUM
-
-
-class PID_Control {
-    public:
-        PID_Control(EepromSettings *nvm, ControlHeat *heat, ControlPWM *vent, std::function<float(uint8_t)> getLogicalChT):
-            _nvm(nvm),
-            _heat(heat),
-            _vent(vent),
-            getLogicalChanTemp(getLogicalChT) {};
-        bool begin();
-        void loadProfile( uint8_t profileNum );
-        bool isOn();
-        void turnOff();
-        void turnOn();
-        void updateCycleTimeMs( uint32_t ct );
-        void updateSetPointC( float setPointC );
-        float getSetPoint() { return this->setp; };
-        void updateTuning( float kP, float kI, float kD );
-    
-    protected:
-        EepromSettings      *_nvm;
-        ControlHeat         *_heat;
-        ControlPWM          *_vent;
-        std::function<float(uint8_t)> getLogicalChanTemp;
-        HardwareTimer       *_timer;
-        bool                isInitialized = false;
-        float               input = 0;
-        float               output = 0;
-        float               setp = 0;
-        QuickPID::Control   _action = QuickPID::Control::manual;
-        QuickPID            _pid = QuickPID(&input, &output, &setp);
-        void _compute();
-};
-
 
 class StateCommanded {
     public:
